pyproject.toml
[tool.poetry]
name = "ai_campus_admin_agent"
version = "0.1.0"
description = "FastAPI backend with LangGraph agent for campus admin"
authors = ["AI Assistant <assistant@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
uvicorn = { version = "^0.30.6", extras = ["standard"] }
python-dotenv = "^1.0.1"
pydantic = "^2.9.2"
pydantic-settings = "^2.5.2"
motor = "^3.6.0"
email-validator = "^2.2.0"
langchain-core = "^0.3.9"
langchain-openai = "^0.2.4"
langgraph = "^0.2.45"
sse-starlette = "^2.1.2"
httpx = "^0.27.2"
duckduckgo-search = "^6.3.0"
elevenlabs = "^1.6.2"
speechrecognition = "^3.10.4"
pyaudio = "^0.2.14"
python-jose = { version = "^3.3.0", extras = ["cryptography"] }
passlib = { version = "^1.7.4", extras = ["bcrypt"] }
python-multipart = "^0.0.9"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
ruff = "^0.6.9"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"


